requirements.txt
openai==1.23.6
python-telegram-bot==21.1.1